third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGPatternElement-svgdom-width-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'width' property of the SVGPatternElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS patternElement.width.baseVal.value is 100
PASS patternElement.width.baseVal.value is 50
PASS successfullyParsed is true

TEST COMPLETE
